Extracts reusable CoreData helper functions into separate file.
